backend/requirements.txt
Flask
python-dotenv
google-generativeai
PyPDF2
Werkzeug


opencv-python 
mediapipe 
numpy

Flask
pandas
PyPDF2
nltk
fitz 
pypdf
werkzeug
google-generativeai
pyyaml
logging
jinja2


Flask
python-dotenv
google-generativeai
PyPDF2
Werkzeug


fastapi
uvicorn
google-generativeai
whisper
torch
soundfile
python-dotenv
kokoro
opencv-python
mediapipe
deepface
fer
numpy
pandas
matplotlib
scipy
